move most things out of echox
