+++ internal/ent/mixin/audit_mixin.go
@@ -8,7 +8,7 @@ import (
 	"entgo.io/ent/schema/field"
 	"entgo.io/ent/schema/mixin"
 
-	"github.com/datumforge/datum/internal/echox"
+	"github.com/datumforge/datum/internal/auth"
 )
 
 // AuditMixin provides auditing for all records where enabled. The created_at, created_by, updated_at, and updated_by records are automatically populated when this mixin is enabled.
@@ -59,7 +59,7 @@ func AuditHook(next ent.Mutator) ent.Mutator {
 			return nil, newUnexpectedAuditError(m)
 		}
 
-		actor, err := echox.GetUserIDFromContext(ctx)
+		actor, err := auth.GetUserIDFromContext(ctx)
 		if err != nil {
 			actor = "unknown"
 		}
